Add files via upload
Basically i am creating prompts and importing argv from sys to help me with my input. The prompt is where the variables are saving that the user inputs i am guessing. Just read through this exercise haven't had much of a chance to play with it yet. But will fuck around with it more tomorrow, off monday & tuesday so i will have way more time to play around with it.