Add files via upload
So in my book the author assigned me a task to create a text based decision based game. Basically introducing me to else, elif & if statements. So it works but theres a logical error and i can't figure out why, once you hit the end of the first part of it it dosen't expand but needs to expand to create further choices. I put some comments in the code hoping it helps, i really appreciate you taking the time to do this for me.