Add files via upload
So in my book right now the author is asking me to create a game using else if elif statements, he wants me to make it as complex as possible chaining together elif else and if statements/branch them out as much as possible. So the problem is after you choose option 1 and go through everything it ends. There should be another branch that continues off option 1, so if you chose that option it leaves on a "cliff hanger" then you make three more decisions based off option 1. But it stops before it can branch out more if that makes sense. He has example code that i tried to follow but could not seem to figure out how to continue branching.